Move EnEx transition sync to player packets
