+++ client/src/CEntryExitTransitionSync.h
@@ -3,7 +3,7 @@
 class CEntryExit;
 class CPed;
 
-namespace Packets::Scripts
+namespace Packets::Players
 {
 class EnExTransition;
 }
@@ -13,7 +13,7 @@ class CEntryExitTransitionSync
 public:
     static void OnTransitionStarted(CEntryExit* entryExit, CPed* ped);
     static void OnTransitionFinished(CEntryExit* entryExit, CPed* ped);
-    static void Receive(const Packets::Scripts::EnExTransition& packet);
+    static void Receive(const Packets::Players::EnExTransition& packet);
     static void Process();
 
 private:
